+++ ICD.Profound.ConnectPRO/Rooms/ConnectProRoom.cs
@@ -75,10 +75,19 @@ namespace ICD.Profound.ConnectPRO.Rooms
 		public IVolumeDeviceControl GetVolumeControl()
 		{
 			IVolumePoint volumePoint = Originators.GetInstance<IVolumePoint>();
+			if (volumePoint == null)
+				return null;
 
-			return volumePoint == null
-				       ? null
-					   : Core.GetControl<IVolumeDeviceControl>(volumePoint.DeviceId, volumePoint.ControlId);
+			try
+			{
+				return Core.GetControl<IVolumeDeviceControl>(volumePoint.DeviceId, volumePoint.ControlId);
+			}
+			catch (Exception)
+			{
+				Log(eSeverity.Error, "Failed to find volume control for {0}", volumePoint);
+			}
+
+			return null;
 		}
 
 		/// <summary>
@@ -221,16 +230,18 @@ namespace ICD.Profound.ConnectPRO.Rooms
 			try
 			{
 				if (string.IsNullOrEmpty(dialingPlanPath))
-					Logger.AddEntry(eSeverity.Warning, "{0} has no Dialing Plan configured", this);
+					Log(eSeverity.Warning, "No Dialing Plan configured");
 				else
 				{
-					string dialingPlanXml = IcdFile.ReadToEnd(dialingPlanPath, Encoding.ASCII);
-					m_ConferenceManager.DialingPlan.LoadMatchersFromXml(dialingPlanXml);
+					string xml = IcdFile.ReadToEnd(dialingPlanPath, new UTF8Encoding(false));
+					xml = EncodingUtils.StripUtf8Bom(xml);
+
+					m_ConferenceManager.DialingPlan.LoadMatchersFromXml(xml);
 				}
 			}
 			catch (Exception e)
 			{
-				Logger.AddEntry(eSeverity.Error, "{0} failed to load Dialing Plan {1} - {2}", this, dialingPlanPath, e.Message);
+				Log(eSeverity.Error, "failed to load Dialing Plan {0} - {1}", dialingPlanPath, e.Message);
 			}
 
 			// If there are no audio or video providers, search the available controls
@@ -268,7 +279,7 @@ namespace ICD.Profound.ConnectPRO.Rooms
 			}
 			catch (Exception e)
 			{
-				Logger.AddEntry(eSeverity.Error, "{0} failed add {1} dialing provider - {2}", this, sourceType, e.Message);
+				Log(eSeverity.Error, "failed add {0} dialing provider - {1}", sourceType, e.Message);
 			}
 		}