+++ ICD.Profound.ConnectPRO/Themes/UserInterface/IPresenters/Common/IVolumePresenter.cs
@@ -1,5 +1,4 @@
 ﻿using ICD.Connect.Audio.Controls;
-using ICD.Connect.Devices.Controls;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common;
 
 namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.Common